requirements.txt
python-dotenv==1.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.36
langchain-groq==0.1.6
sentence-transformers==3.0.1
huggingface-hub==0.24.5
chromadb==0.5.5
pypdf==4.3.1
crewai==0.36.1
crewai-tools==0.13.0
tqdm==4.66.5
requests==2.32.3
pydantic==2.12.5
openai==0.27.14
